Suppressed some warnings in Shadow-Detection module. Added ability to pass a string argument to test benches. Added new widget for marking and editing ground control points on simulation data sets. New organizational scheme for simulation data sets. Now loaded directly from Simulation-Data-Sets without needing to be copied into BIN. The previous system had too many files and some things were still hard-coded (like where the ref frame comes from). Now the ref frame is loaded from a JPEG in the data set folder, and the GCPs (pixel coords and Lat/Lon/Alt) are loaded from GCPs.txt, which is parsable but also easily human-editable. Much lower risk of mistakes (like mixing ref frames and GCP files) which could result in messed up fiducials. Updated testbenches to support argument passing and updated all testbenches that use SimulatedDrone to use new scheme for loading data sets. New utilities for string handling added for supporting data parsing and testbenches.

+++ SRC/Modules/Shadow-Detection/shadow_utils.hpp
@@ -98,8 +98,8 @@ inline void getApertureMask(const cv::Mat& img, cv::Mat& mask) {
 // Value is supposed to be CV_8UC3 but only the first value is used
 inline void create_masked_binary(const int mapRows, const int mapCols, const cv::Mat mask, cv::Mat binary_sampled){
 
-    for (uint32_t row = 0U; row < mapRows; row++) {
-        for (uint32_t col = 0U; col < mapCols; col++) {
+    for (uint32_t row = 0U; row < (uint32_t) mapRows; row++) {
+        for (uint32_t col = 0U; col < (uint32_t) mapCols; col++) {
 	    // Set any pixel with value of 254 to 255
 	    if (binary_sampled.at<uchar>(row,col) == 255){
 		//std::cout << "Settting value at " << row << "," << col << " to 254" << std::endl;
